move nlohmann all to rpc
