+++ src/blockchain_db/CMakeLists.txt
@@ -39,7 +39,6 @@ target_link_libraries(blockchain_db
     lmdb
     filesystem
     Boost::thread
-    nlohmann_json::nlohmann_json
     extra)
 
 target_compile_definitions(blockchain_db PRIVATE
