+++ src/checkpoints/CMakeLists.txt
@@ -37,6 +37,5 @@ target_link_libraries(checkpoints
     cryptonote_basic
     Boost::program_options
     Boost::serialization
-    nlohmann_json::nlohmann_json
     filesystem
     extra)
