+++ src/cryptonote_basic/cryptonote_basic.h
@@ -45,7 +45,6 @@
 #include "ringct/rctTypes.h"
 #include "device/device.hpp"
 #include "txtypes.h"
-#include <nlohmann/json.hpp>
 
 namespace service_nodes
 {
@@ -65,15 +64,6 @@ namespace service_nodes
       FIELD(voter_index)
       FIELD(signature)
     END_SERIALIZE()
-  };
-
-  inline void to_json(nlohmann::json& j, const quorum_signature& s)
-  {
-    j = nlohmann::json
-    {
-      {"voter_index", s.voter_index},
-      {"signature", tools::type_to_hex(s.signature)},
-    };
   };
 };
