+++ src/cryptonote_core/service_node_list.h
@@ -39,7 +39,6 @@
 #include "cryptonote_core/service_node_quorum_cop.h"
 #include "common/util.h"
 #include "uptime_proof.h"
-#include <nlohmann/json.hpp>
 
 namespace cryptonote
 {
@@ -373,7 +372,6 @@ namespace service_nodes
     END_SERIALIZE()
   };
 
-  inline void to_json(nlohmann::json& j, const key_image_blacklist_entry& b) { j = nlohmann::json{{"key_image", tools::type_to_hex(b.key_image)}, {"unlock_height", b.unlock_height}, {"amount", b.amount} }; };
 
   struct payout_entry
   {
