+++ src/multisig/CMakeLists.txt
@@ -35,5 +35,4 @@ target_link_libraries(multisig
     ringct
     cryptonote_basic
     common
-    nlohmann_json::nlohmann_json
     extra)
